move to folder
moved files to a dedicated constants folder and made necessary refactoring changes
