+++ src/main/java/frc/robot/RobotState.java
@@ -4,6 +4,7 @@ import edu.wpi.first.math.Matrix;
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.numbers.N2;
 import edu.wpi.first.math.numbers.N3;
+import frc.robot.Constants.ConstantsBase;
 
 public class RobotState {
     private ConstantsBase constants;
